[32179] Reload list when deleting work package
We have the halEventsService to notify other (active!) views about changes to resources such
as deletion.

However in the full view, the list view is destroyed and does not listen
to these changes, thus going back to the list after deletion with a
loaded query does nothing currently and renders the old list.

The easiest solution for this is to simply reload the list in this case
through ui-router.

https://community.openproject.com/wp/32179