display user name as link for non admins
The showUser link of the user representer is now used to signal when the link can be followed. That link can only be followed when the user is not logged. The frontend now only checks the availability of the link to determine whether the user name in the work package activity is to be displayed as a link or not