[31628] Downcast todo lists in tables to markdown
Tables are saved as HTML, thus avoids any frontend turndown within it.
This means todo lists are kept as is, with the CKEDitor frontend style.
Currently, they will be sanitized after turned to markdown in the
backend.

We can easily downcast the todo list part into text.
This ensures that the todo list does not get sanitized, and will
correctly be found by the todo list filter.

https://community.openproject.com/wp/31628