reimplement direct jump to wp on search hit
If the user presses enter before the search results are back in, the component now waits until the results are back and then decides on what to do (go to
work package on direct hit or trigger search otherwise.)

This almost recreates the behaviour we had before with the only difference from the user perspective, that the user might have to wait a tad longer and the result list is show briefly.

This is way simpler code than what we had before in the backend and I believe the drawback to be negligible.