deepClone payload source object to avoid side effects
Every change to the source would otherwise alter the source of the form/resource when the intend of the method is to build a payload. In the case of time entries, the project is removed as it will later be deduced by the work package. But if the project reference in the source is removed, later code will fail when trying to instantiate the project (which is done lazily).