allow fetching activities by an id referencing a subsumed journal
AggregatedJournals might contain a set of native journals. This change
allows fetching an aggregated journal the id of any journal included in
the aggregated.

This provided clients with more available references to an aggregated
journal. This is particularly helpfull in case the aggregated journal
includes more journals than the client expects. This might happen when
the journable is updated in the meantime. E.g. a user adds a screenshot
to the last comment within the time frame available for aggregation; a
user adds a comment, alters an attribute and wants to change the
comment.

At the same time, this change limits the aggregated journals created in
the database to a particular journable. This has a severe performance
impact as without the focus on a particular journable, potentially
millions of journals will be aggregated to then only fetch one
individual result.