Fix cukes in backlogs due to js cucumber
We now need to run cukes with a JS scenario all the time. Backlogs
however used manually POST requests to set up some stuff which is no
longer working now.

We can tell capybara to switch the driver shortly just to perform the
request while keeping the session. It's a hack but will prevail until
backlogs is hopefully removed soon.