[31617] Re-enable group sums rendering
The grouped rows builder excluded rendering of per-groups sums,
however the API returned these results anyway.

It's easy to just render a sums row as we did before. To that end,
I have removed that functionality from the sums row directive and into
a generic renderer that can be used with and without angular.

https://community.openproject.com/wp/31671