Add `deployUrl` for cloud
Importing relative assets from components will get them output to the
deployPath and prepended with the `deployUrl`.

This unfortunately is static, and not supported with our
RELATIVE_URL_ROOT rails option, which currently does not trigger
compilation.

On cloud, this is not a problem so enabling this option for the
subscription plugin