Flexible widget title input field width
When entering a text, the input field now autoresizes with the input text. This gracefully handles overflowing, but
requires browsers to support CSS grid, which all of the browsers in browserslistrc do.

This solution is a bit hacky on the CSS side, but requires almost no javascript to function, and utilizes the browser layouting engine to do width calculations. This is both faster and more accurate, since it accounts for the specific widths of the font being rendered, as well as sub-pixel sizes.