remove operators not applicable to only subprojects filter
The none operator does not make sense as it means that no work package will be returned. And the all operator will be the same as not having the filter at all.

There is a minor change in the UI in this change in that the order of the opertors of the subproject filter is altered. But this should not be a problem for the user and leads to a simpler implementation.