add derived dates to the work package resource
The derived dates are the minimum/maximum of the descendant's dates. As
long as the work package is scheduled automatically, the derived and non
derived dates should be equal althoug the derived dates will be null in
case non descendants exists. But if the user decides to activate manual
scheduling the dates can deviate. The derived dates can then help to
display the deviation to the user.

This commit however does not alter the front end.

The dates are eager loaded on the work package index commit using the
same mechanims already introduced for spent time. The patched cost
attributes, which are also eager loaded, are now eager loaded in the
core as well so the patch is removed. As those attributes are now part
of the core (although in a different module) it no longer makes sense to
rely on patching for this.