Flexible widget title input field width (#8516)
* Flexible widget title input field width

When entering a text, the input field now autoresizes with the input text. This gracefully handles overflowing, but
requires browsers to support CSS grid, which all of the browsers in browserslistrc do.

This solution is a bit hacky on the CSS side, but requires almost no javascript to function, and utilizes the browser layouting engine to do width calculations. This is both faster and more accurate, since it accounts for the specific widths of the font being rendered, as well as sub-pixel sizes.

* Fixed white-space issue, added outline while editing for editable-toolbar-title

This Fixed an issue where end-of-input whitespace wasn't rendered in the input shim, causing the input field to not
expand. It also adds an outline around the input when focused.

* Fix editable title shim text

The shim text CSS was too specific to the overview page, the new CSS is less specific, taking override rules the same
way the input does.

* Fixed title-container max-width

After removing the overflow: hidden attribute, the title width could be larger than the parent element. This fixes that.

* Fix unexplicabel text rendering differences.

Adding workaround for unexplicable differences in
text rendering between input and input-shim.
Leading to wrong rendering ellipsis in input-shim.

Co-authored-by: Wieland Lindenthal <w.lindenthal@forkmerge.com>