[34001] Ensure filters always receive the query context
In the mentioned project, the query copy service fails to execute
because a filter has lost its context right after `query.save`.

There is a hook for `after_commit` which is not relevant here since
we're in a transaction, but the context is lost regardless.

Instead, we could ensure whenever accessing the filters, the context
is set correctly.

https://community.openproject.com/wp/34001