Fix/child scheduling restrictions (#8584)
* allow free scheduling children of manualy scheduled parents

If a parent has a follows relations with another work package, that relationship also limites children of the work package. But if the work package is scheduled manually, this creates a scheduling barrier. Children of a manually scheduled work package are not hampered by the follows relationships of the parent or other ancestors.

* simplify soonest_start query

* eager load

* extract to scope