[34000] Ensure user custom fields can be copied
User CFs were not copiable because they depend on project members being
copied.

They need to be overridable in the copy form however, so we need to be
able to set this before and after members are copied. The only option is
to allow visible members to be selected there.

https://community.openproject.com/wp/34000