[33383] Treat number fields as text fields to allow localized input
The budget parsing use localized numbers to compute the budget costs for
the given units.

For comma-delimited locales like German, this will break the computation
for any decimal value, as the frontend number always sends dot-delimited
values in Chrome due to Bug OP#32232

This PR treats the number fields like text fields with a hint on how to
enter the correct value. This matches the "rate" field that does this
already and should be the expected behavior, so users can use their
language for numbers.

https://community.openproject.com/wp/33383