fix touching the customizable on removing custom_values
Before the commit, the existing custom_values where attempted to be
reused but there was no deterministic behaviour with which reusing them
was attempted. Additionally, the existing values where always changed
(or they would receive the current value again) whenever values where
provided, regardless of whether those values actually changes. There
was no explicit flag that a value was removed.

This poses a couple of problems:
* Even if no new values where provided the order of the custom values as
  stored in the db might change leading to journal entries that where
  not actually caused by the user.
* If values where only removed, due to the missing
  explicit removal flag, it became luck whether an existing custom
  value was actually changed. Only then would the customized object be
  touched leading to an update of cached fragments, i.e. representers.
* Because the values where updated all the time, the performance would
  suffer.
* Removed custom values where removed right away and not only after
  save.

This commit introduces an explicit flag for removal and does not alter
existing custom values. This might have the drawback of creating a
custom value even when a value, chosen to be removed by the user, could
be reused but this should be negligible.

It also has the drawback of introducing a flag to be maintained.