Fix: Set current timestamps as default values for BCF timestamps.
The fix default values broke the ActiveRecord timestamp logic. AR did
not set those values with the real cration date. Instead it left
the default values that were defined on DB level. Now, we set the
default values still on the DB level, but with the current
time stamp. Having set a current time stamp as default value comes in
handy, when we need to do some DB maintenance or bulk imports.