set the default value for newly build custom values
Most of the time, the getter #custom_field_values is used with unbeknown to the caller will build custom values for not covered custom values. This method
always assigned nil to the value of the newly build custom value when it should have used the default value. Once that is fixed, all models including acts_as_customizable
will again have set the default values for custom fields correctly. This is true regardless of wether the custom value is added before or after the creation of the model.

Only for work_packages is the additional method #set_default_values! required as the availble_custom_field might change during a request upon a switch in project and/or type.