ensure users are added as new members before groups
That way, conflicts between a group containing an also added user and that user are prevented.

As a user cannot be member of a project twice, an error is produced if the group will add the user as a member and then the user is added in its own right. On the other hand, a user may first be added as member and then a group containing the user may also be added. By simply sorting by group vs. user we can ensure that users are always added first.