Fix #34482: Sometimes no "+ Viewpoint" button present
Especially on the first start of the Add-in this
timing issue became apparent, when the "+ Viewpoint"
button did not show up.

I removed the onLoad of all viewer bridges and
replaced it with a `viewerVisible$` BehaviorSubject.
This way we can asyncronously react to the event
that the viewer becomes "ready".