Fix: Logo was missing when no CustomStyle was set.
This PR extracts the inline CSS for setting the
logo paths.

In the earlier version the there was no default
path for the logo set unless you had a CustomStyle
set.

Apparently, when we did the refactoring of the
global styles and moved the setting of CSS variables
to be completely outside of sprockets, we missed the
logo path.

To test it: Remove all CustomStyle entries from your
DB.