Added PROXY_HOSTNAME env variable for local docker development
Before, the webpack dev proxy tried to proxy any requests to `localhost:3000`. However, since the backend and frontend
containers are different hosts in the docker context, it was proxying to itself and failing. The proxy now takes this
env variable which is set in the docker-compose file to `backend`, but defaults to localhost for backward compatability.