Fix broken call to #super in Budget#copy_from
The #copy_from method in Budget appears to assume that superclasses
will have the same method available to them as called on the
new Budget object to which the copy is going. This is not the case.

Resolve the issue by assigning duplicated attributes from the
source object directly to self.attributes prior to copying the
actual budget items.

Tested in Chef/Vagrant test environment.
Deployed to SemperVictus pre-production OP instance successfully.