Fix calendar widget resize when adding or removing other widgets (#8784)
* Fix calendar widget resize when adding or removing other widgets

The fullcalendar dependency only reacts to the `document.resize` event. However in our case the widget can also resize
when neighboring widgets are added or removed. This fix adds a `ResizeObserver` listener to account for that. It is
supported in all evergreen browsers.

* Explicitly add resize observer typings

The typings file for ResizeObserver was not being autoloaded. This commit adds the typings explicitly, making sure tsc
picks up on them.

* Move typings include to a different file