Feature: Optimize memory for pdf exports with attachments (#8861)
* Change design of WP list PDF export to batches

In order to keep memory consumption low when
exporting WP as list in PDF, we split the work
in batches of 100 work packages. In every batch
we resize the attached photos and create a PDF.
When all batches have been processed, we merge
the PDF files via the command line tool
`pdfunite` which fortunately comes with a package
that we already require for text extraction.
Only at the end, when the merged
PDF is ready (and small) we load it into the
ruby process to return the PDF's content with
service's success.

In every batch, we make sure that we remove
all minimized images afterwards. And after
merging the batch PDF files to the final
PDF we also delete all batch PDFs.

* Don't load the merged PDF file into the ruby process

* Rename tempfile to final file name before storing.

* Improve code structure.