Fix missing link on single cards
Work Package cards have a link to their ID to the full view, which was
handled by ui-router srefs. This was replaced in https://github.com/opf/openproject/pull/8544,
leaving the link without href to use with modifiers.

This PR restores the functionality of using modifiers to copy the
link/open in a new tab