prevent css styles of form--section to affect table (#8926)
In the summary page, the report tables are wrapped within a div with the form--section class even though those reports are not actually part of a from. Ideally, a new class would be styled containing only the desired css rules. But as the view is deemed to be replaced anyway, the quick and dirty way is to simply take the table out of the form--field's dom structure