Fix/journals on direct uploads (#8933)
* linting

* use Carrierwave's replacement for filename

This avoids having the name altered by Carrierwave later on in the Attachments::FinishDirectUploadJob job where the file is fetched and then the attachment is stored.

* separate cases for no attachment and no local file

* correct journalizing on direct upload

In case the container of an attachment is journalized, the container needs to have added a journal to it when an attachment is uploaded. Without that, the attachment is only picked up in the journals once the container is altered again (possibly by a different user). This will then lead to a history which does not reflect the actual upload.