Fix/parent filter (#9032)
* use the backend`s url for fetching available values

The backend in it`s schema already provides the valid values. The front end uses this to created a collection resource and fetchh the values from there.

* remove configuration options

For now, the filters are only used in the context of work packages. This might change but having configuration for static values only results in unnecessary complexity. Those options can easily be readded when needed

* correctly blacklist subjectOrId filter

* allow id filter in the front end

Because the autocompleter for work packages is in place now for work package filters, it is trivial to enable the id filter as well