Set moment locale globally
This sets the moment locale to the application locale globally, ensuring that moment uses the right translation strings
wherever applicable. However, since we only work with languages (en), and not real locales (en-US), the first day of
year calculation has to default to some option. This commit defaults to ISO 8601 for every language and locale.