rework changed_by_system (#9043)
Before, it was implemented by passing the changed attribut keys over to
the contract to whitelist them.

This lead to:
  * The contract interface becoming bloated
  * Having to rely on the knowledge of the developer not to falsely
    whitelist an attribute. The developer would also have to make sure
    to not perform a mass assignment after the attribute has been
    whitelisted

The new approach it to integrate the behaviour into the model which is
first altered in the service before it is scrutinized in the contract.

The information about the changed attributes is now stored inside the
model which removes the necessity to flag the whitelisted attribute
separately. Additionally, the exact change is tracked. So if an
attribute is set to one value inside a whitelisted block there is no
risk in later on performing a mass assignment.

This comes at the cost of extending the models which is weird also it is
build into the default SetAttributesService so child classes do not have
to worry. One might include the module into every AR model but currently
we only need it for a very specific use case.