[36269] Add "other" to pluralization rules for one
For languages that do not have "one" as a pluralization rule (japanese,
for example), the pluralization rule defined by I18n.js does not return
a valid string for count: 1.

Since we probably do not want to define custom pluralization rules for
all languages, we can fall back to the "other" key if "one" is not
defined.

https://community.openproject.com/wp/36269