[36764] Fix margins of table when sums and/or inline create showing
In https://github.com/opf/openproject/pull/8771, a double styling was
added when the inline create row is showing. Back then, sums row was
always taking up space even if not visible so the double margin was
correct.

This however is no longer the case. We need to apply different margins
when either sums or inline create, or both are showing.

https://community.openproject.org/wp/36764