fix duplicates returned for wp list on multiple matching attachments/comments
LEFT JOIN leads to potentially having multiple records returned if a search string is present in more than one attachment/comment. This is not a problem if we fetch work packages directly out of the query results but becomes problematic if only the id is fetched and then used to get the work packages by a subquery.