Forbid Factory.build(:user, member_in_project)
If you use the trait member_in_project(s), the user is implicitly saved
to create the member.

This is very confusing if trying to use required custom fields, as this
will fail with the Member#user_id foreign key being nil, as the user
cannot be saved.

Instead, raise an error when trying to use this factory trait