remove reply from message journaling (#9293)
The columns:
* last_reply_id
* replies_count

are columns that exist for performance reasons only. They do not contain information about the message itself.

If they are journaled, adding a message to a forum updates the last_reply_id of the parent message. The user replying will then be shown to have updated the replied to message