[#37440] OpenProject backup fails on etc folder
The backup script was changed to include the config folder in
https://github.com/opf/openproject/pull/9189

This however results in permission errors depending on how the apache
integration is configured. Files added there are owned by root, not
openproject.

The reason behind the PR was to backup the installer.dat and conf.d
entries however, so we can just explicitly add those, avoiding the
permission errors.

https://community.openproject.org/work_packages/37440