enforces dependent project modules being active
Project modules, e.g. calendar can define dependencies, e.g. work packages. An error message is now displayed if a user enables a project module without the dependency being active in the project as well.

Also fixes a serious bug in the Shared::ServiceContext module. Whenever a failing service result is returned, every change to the database is to be rolled back. This was not true when a model was provided since the changes where carried out within an inner transaction