ensure project status is saved (#9325)
aweseome_nested_set, used for maintaining the project hierarchy, reloads
the project after saving. Any change to an associated model, e.g. the
status, is lost upon reloading.

Thus, saving of the status needs to take place before saving the
project.

This is safe to do since it is carried out in a transaction and all the
validations are run for the status as well.

https://community.openproject.org/work_packages/37464