Fix dropdown size for ium project and user dropdowns (#9372)
The dropdowns are using custom disabled elements, which are taller than the rest. This causes virtualscroll to miscalculate the needed height for the dropdown, resulting in a scrollbar. The easy fix is to disable virtualscroll. This makes rendering of large lists a tad slower, but this can be mitigated by e.g. only searching after the second character has been typed should it be necessary.