Fix how we gather and derive sampling traces for sentry
We previously defined a fixed tracing rate AND a sampling tracer in the backend,
which result in everything to be traced regardless of the actual rate.

We now define a base rate for certain requests, so that we can exclude health checks and other things.
The configuration is now a factor (0 = off, default) (1 = use suggested rates), any other value x (x multiplied by suggested rates)