Load full HAL collection resources (#9383)
HAL Collections only load the first pages when `$load` gets called. This means that parts of the application that rely on the full list being available fail. One example of this are the multi select filters, specifically for users; only one page gets loaded initially, so when searching for users you don't see the full list.

On large instances, this could result in long loading times. Though the calls are fully async and the UI will not block, this might surprise some users since we do not show a loading indicator.

A better fix would be to rewrite the filter-toggled-multiselect-value component, but after an initial inspection this
would require a ton of work.

Long term, we're looking at removing HAL classes altogether, so if the loading times are reasonable this could be a
reasonable workaround.

Co-authored-by: Oliver Günther <mail@oliverguenther.de>