exclude full text columns from attachment model
Currently, the fulltext columns `fulltext`, `fulltext_tsv` and
`file_tsv` are never displayed in the application.  As such they are
currently purely used internally when the are involved in the fulltext
search. Having them included in the AR model Attachment will lead to a
lot of data being loaded. Such a column can contain megabytes of data.
This is costly twice: when fetching the data from the database and when
instantiating the model in ruby.  As the data is never displayed, they
can be ignored for now using `ignored_columns`. This changes the SQL for
fetching from

`SELECT "attachments"."*" ...`

to

`SELECT "attachments"."id", "attachments"."file" ...` (without the excluded columns)

The alternative would have been to create a new table for the fulltext
information and this might still be an option once the requirements on
the columns increase. For now, this change is simpler.