avoid creating a journal for a to be destroyed work package
Work packages with attachments where saved once for every
attachment upon destruction. This led to custom fields being rewritten
and a journal being created. Those created records where removed right
away, since the work package was destroyed so the problem was only
perceivable as poor performance.

On the other hand, we had reports of users not being able to destroy
work packages. While it is pure speculation that this change fixes the
problem it at least makes deadlocks less likely since journaling is
performed with a mutex and that no longer takes place now.

The `after_remove` callback is no longer necessary as journaling is
triggered in the services now.