[38325] Using existing gem for email validations
Our email validation is too strict for RFC compliant emails.
Instead of trying to do it ourselves (email validation regexps are hard!),
let's just use a gem to do that for us.

Note: The gem also has a strict mode that could be optionally enabled. This however
results in some emails on community to become invalid. I left it on the default mode for now.

https://community.openproject.org/work_packages/38325