introduce resourceId and resourceType filter
These filters in combination allow filtering notifications via the
associated resource, e.g. a work package. When only using the id filter,
due to the polymorphic association of the resource, it will not be
guaranteed to only return the desired notifications. A different
resource, e.g. a meeting might have the same id.

But using the resourceType filter by itself also makes sense so that a
client can fetch only notifications associated to a subset of resources,
e.g. work packages.