prevent author watcher on work package copy
Newly created work packages get their author assigned as a watcher. When copying, this would mean that the copying user is now watcher in all of the copied work packages. This commit changes the behaviour to only include the user as a watcher if that user was also watching the source work package.