increase robustness of journable_type migration
In case a journal with a journable_type of `Budget` already exists, that journal will have a version that will later on conflict with the unique index on
(journable_id, journable_type, version)
when renaming all journable_type columns from `CostObject` to `Budget`. To prevent the problem, the version of already existing journals with `Budget` are increased by the highest version of their `CostObject` counterparts.