Rework unread counts for notification stores
So far, notification stores held two different types of information; notifications themselves, and an "unread count",
which was fetched separately for the notification bell icon. However, the new stores are meant to fetch the filtered
lists directly. This commit removes the unread count code from stores, and introduces the ability to set page sizes,
allowing the bell icon to fetch an empty page and read the unread count from the total number returned by the backend.

Also fixes a double subscribe inside the notifications service on fetch.