change rounding scale from 2 to 4 digits
A rate will be saved with a scale of 4 in the database.
Without this patch the form is prefill with a rounded rate with a scale of 2 when you open the update form even when a scale of 4 is defined in the database.
You won't see the real value of the database. Even worst is when you click on save you will save this new rounded value.
Example:
You define a rate of 0.125 and save it. In the database the 0.125 is saved.
When you reopen the rates form you will see an 0.13.
When you open the update view of the rate, you will see a prefilled form with 0.13.
When you now click on save the untouched value will be saved to the database.
In the database the rate of 0.13 saved.

This patch change the update view so that the prefilled form shows the rate with a scale of 4 like it is saved in the database.