[38813] Ensure embedded tables load their own columns
Previously, if `showFilters` was false, the embedded tables would not
load their own query form but piggyback on the globally loaded form.

This however does not work when:

1. Opening a work package in a deep link (global query form never
loaded)

2. The embedded table resides in a subproject (different query form
necessary)

That's why this PR moves the available states into the isolated query
space and ensures the form is always loaded.

https://community.openproject.org/wp/38813