add simple eager loading to the notifications api
Unfortunatly, the `data` association cannot be eager loaded since it is a polymorphic association and the successor journal (including it`s data, customizable_journal and attachable_journal) cannot be eager loaded since there is no association for it at all. Some eager loading wrapper similar to what has been implemented for work packages would help here. It could then also be applied to the activities of an individual work package