eager loading for notifications and activities
Associations that cannot be loaded via rails` eager loading are eager
loaded by hand:
* The 'data' association of journals
* The 'predecessor' association of journals
* The 'journable' association of journals
* The 'resource' association of notifiations

Since the notifications and the activities index api have a similar data
structure, notifications can reuse the eager loading of activities.

Does not include any optimization for the rendering of the details as of
now since the structure of the details is still subject to change.