[38842] Route back to notifications when error happens
Instead of showing an empty split view, route back to the base route
and set a flash message to be rendered after the route completes.

For that to happen, we need a common root state that we currently did
not have. This PR thus has to change some of the route definitions and
uses views to set the parent ui-view instead of creating a new
component. This reduces the weird stacking effect of the ui-views
currently present.

https://community.openproject.org/wp/38842