Also reload activities in the work package split view on update, not just the notification count
When there are new notifications in the center, the user can force an update to the list. We also updated the number of
notifications inside the work package, but did not load the new activity tab entries. This is fixed in this PR.

Relevant comment:

https://community.openproject.org/work_packages/39466/activity#activity-6

While we're here, also change the wp-split-view selector to align with the filename