[38606] Ensure we always use the english slug
When creating wiki pages, `to_url` uses the I18n.locale used on startup,
which is set to the default_language setting.

We don't want that, and instead always want to create english slugs for
consistency.

To still find pages created with the default_language flag, `find_page`
is extended to fall back to the later one.

https://community.openproject.org/wp/38606