Proof of Concept for a DatasetInputs decorator
Currently, there are some usecases where we render a component tag in the backend, and then hydrate it in the browser.
Accepting inputs in this scenario is less than preferable since the checking and parsing for dataset inputs has to be written
in the component itself.

This decorator does this automatically in the background. Just add it to a component, and it checks automatically
whether there are data attributes to set initial input values from.

In order not to overwrite any class properties that are not inputs, I've added a bit of a hacky check to get the
Angulars internal `declaredInputs` object, which, as the name implies, lists the declared inputs. This is improvable.