avoid updating/initializing the query space twice initially
Both the `PartitionedQuerySpacePageComponent` as well as the
`WorkPackagesListService` initialize the work package states via the
`WorkPackageStatesInitializationService`.

The `PartitionedQuerySpacePageComponent` initializes the
`WorkPackagesListService` so the former is indirectly for calling both.

Ideally, only the `WorkPackagesListService` would initialize the
`WorkPackageStatesInitializationService` to avoid confusion but this
could not be implemented at this time because of hard to control side
effects.

With this change, the `PartitionedQuerySpacePageComponent` on the first
call to the `WorkPackagesListService` no longer initializes the wp
states trusting in the list service to do it.

This change results in reduced initialization time and avoids duplicated
requests to the server e.g. for relations and the work packgages they
point to.