Precompute moment dates for activities
The current activity info class dynamically constructs moment objects each time the `date` or `dateOfPrevious`
properties are accessed. When rendering the list, this causes the date objects for all activities to be computed at
least twice. A mouseenter or mouseleave generates a rerender in this context. This would normally not be an issue, but
when scrolling inside a long list of activities this causes huge render times and fps drops.

This commit changes the date computation; doing it once when the activities are fetched, so that subsequent renders of
the list will be faster.