Attempts to improve the "attachments" models.
There's lots of redundancy in the API spec which leads to generation of
duplicate but distinct / incompatible model classes during code
generation.

The "attachment" object type is only on of several places where this
occurs in the spec, but I'd get feedback if this unification is desired
first before investing more time in cleaning up definitions.