Improve `watchers_model.yml` and `work_packages_model.yml`
Open question: `work_packages_by_project_model.yml` is structurally
identical to `work_packages_model.yml`, the only difference is the
embedded `example:`.

I can imagine that it's important, due to the focus on API doc
generation from this spec, so one option would be to just reference the
`work_packages_model.yml` for the structural definition, avoiding
redundancy for this, but in this case leaving the redundancy in the
`example:` definition, with all possible drawbacks (inconsistencies
etc.)

Another approach would be to get rid of
`work_packages_by_project_model.yml` completely and just using the
generic `work_packages_model.yml` everywhere where we have a list of
Work Packages.