proper representer initialization to ensure cf inclusion
Representers including the CustomFieldInjector need to use the create method to ensure proper handling of the cf properties. If that representer is cached, as is the case with WorkPackages, not using the .create method will affect every subsequent rendering of the representer as well since most of the information is cached.