Fix error when changing parent of sub-sub-work_package
If the parent of a work package is changed and the new parent has any
common ancestors with the old parent, then the change will fail and the
following message will be shown in the UI:
Could not update the resource because of conflicting modifications.
Click here to refresh the resource and update to the newest version.

This happens because ActiveRecord optimistic locking is enabled and
update_ancestors_service.rb does the following:
* Retrieve all new ancestors and update estimated_hours and done_ratio
  without saving.
* Retrieve all old ancestors and update estimated_hours and done_ratio
  without saving.
* Save both the new and old ancestors.
If there are any work packages that are both new and old ancestors then
those work packages will be retrieved from the database twice, both
copies will be updated, then both copies will be saved, resulting in an
optimistic lock conflict.

This commit corrects this issue by collecting and deduplicating the list
of ancestors before updating them.